refactor(channels): consolidate channel architecture per Issue #163 Phase 3
This PR implements Phase 3 of Issue #163 - consolidating the channel
architecture by moving components to their proper locations.

## Changes

### Core Components (moved to src/core/)
- TaskFlowOrchestrator: Manages dialogue execution phase
- MessageLogger: Persistent message logging to chat-specific MD files

### Feishu Platform Components (moved to src/channels/platforms/feishu/)
- feishu-file-downloader.ts: File download functionality
- feishu-file-uploader.ts: File upload functionality

### Cleanup
- Removed duplicate card-builder files from src/feishu/ (already exist
  in src/channels/platforms/feishu/card-builders/)
- Updated all import paths
- Updated src/feishu/index.ts with @deprecated re-exports for backward
  compatibility
- Updated tests to use new import paths

## Benefits
- Clear separation of platform-agnostic vs platform-specific code
- Better code organization aligned with target architecture
- Maintains backward compatibility through re-exports

Implements #163
