+++ src/channels/feishu-channel.ts
@@ -10,11 +10,11 @@ import { Config } from '../config/index.js';
 import { DEDUPLICATION, REACTIONS } from '../config/constants.js';
 import { createLogger } from '../utils/logger.js';
 import { attachmentManager } from '../core/attachment-manager.js';
-import { downloadFile } from '../feishu/file-downloader.js';
-import { messageLogger } from '../feishu/message-logger.js';
+import { downloadFile } from './platforms/feishu/feishu-file-downloader.js';
+import { messageLogger } from '../core/message-logger.js';
 import { FeishuFileHandler } from './platforms/feishu/feishu-file-handler.js';
 import { FeishuMessageSender } from './platforms/feishu/feishu-message-sender.js';
-import { TaskFlowOrchestrator } from '../feishu/task-flow-orchestrator.js';
+import { TaskFlowOrchestrator } from '../core/task-flow-orchestrator.js';
 import { TaskTracker } from '../utils/task-tracker.js';
 import { BaseChannel } from './base-channel.js';
 import type { FeishuEventData, FeishuMessageEvent } from '../types/platform.js';
