+++ src/core/index.ts
@@ -7,6 +7,8 @@
  * Components:
  * - MessageHistoryManager: Tracks conversation history per chat
  * - AttachmentManager: Manages pending file attachments per chat
+ * - TaskFlowOrchestrator: Manages dialogue execution phase
+ * - MessageLogger: Persistent message logging to chat-specific MD files
  *
  * These components are extracted from Feishu-specific implementations
  * to enable reuse across all channel types.
@@ -24,3 +26,13 @@ export {
   AttachmentManager,
   attachmentManager,
 } from './attachment-manager.js';
+
+export {
+  TaskFlowOrchestrator,
+  type MessageCallbacks,
+} from './task-flow-orchestrator.js';
+
+export {
+  MessageLogger,
+  messageLogger,
+} from './message-logger.js';
